+++ src/packages/__VUE/cascader/demo/index.vue
@@ -18,7 +18,7 @@
 </template>
 
 <script setup lang="ts">
-import { useTranslate } from '../../../utils';
+import { useTranslate } from '@/sites/utils';
 import Basic from './basic.vue';
 import CustomAttributeName from './custom-attribute-name.vue';
 import AsyncLoading from './async-loading.vue';
